feature/update: game object library, MapAssets
update game object library, MapAssets add CustomPlatform

+++ Scripts/GameObjects/Model/GameObjectAddGameObjectAssetModel.cs
@@ -16,6 +16,8 @@ namespace Ursula.GameObjects.View
         {
             AudiosFrom = new List<string>();
             AnimationsFrom = new List<string>();
+            AudiosTo = new List<string>();
+            AnimationsTo = new List<string>();
         }
 
         public string ModelName { get; private set; }
@@ -26,8 +28,12 @@ namespace Ursula.GameObjects.View
         public string ModelPath { get; private set; }
         public string DestPath { get; private set; }
 
-        public List<string> AudiosFrom { get; private set; }
-        public List<string> AnimationsFrom { get; private set; }
+        public string GameObjectGroup { get; private set; }
+        public int GameObjectClass { get; private set; }
+        public string GameObjectSample { get; private set; }
+
+        public List<string> AudiosFrom { get; set; }
+        public List<string> AnimationsFrom { get; set; }
 
         public List<string> AudiosTo { get; set; }
         public List<string> AnimationsTo { get; set; }
@@ -80,12 +86,28 @@ namespace Ursula.GameObjects.View
             PreviewImageFilePath = value;
             return this;
         }
+        public GameObjectUserSourceData SetGameObjectGroup(string value)
+        {
+            GameObjectGroup = value;
+            return this;
+        }
+        public GameObjectUserSourceData SetGameObjectClass(int value)
+        {
+            GameObjectClass = value;
+            return this;
+        }
+        public GameObjectUserSourceData SetGameObjectSample(string value)
+        {
+            GameObjectSample = value;
+            return this;
+        }
     }
 
     public partial class GameObjectAddGameObjectAssetModel : IInjectable
     {
         public bool IsGameObjectAddUserSourceVisible { get; private set; } = false;
         public bool IsEditMode { get; private set; } = false;
+        public string Provider { get; set; } = GameObjectAssetsUserSource.LibId;
 
         public GameObjectAssetInfo assetInfo;
         public GameObjectTemplate template;
@@ -98,6 +120,9 @@ namespace Ursula.GameObjects.View
         public string GraphXmlPathTo => DestPath + Path.GetFileName(_gameObjectUserSourceData.GraphXmlPath);
         public string PreviewImageFilePathFrom => _gameObjectUserSourceData.PreviewImageFilePath;
         public string PreviewImageFilePathTo => DestPath + Path.GetFileName(PreviewImageFilePathFrom);
+        public string GameObjectGroup => _gameObjectUserSourceData.GameObjectGroup;
+        public int GameObjectClass => _gameObjectUserSourceData.GameObjectClass;
+        public string GameObjectSample => _gameObjectUserSourceData.GameObjectSample;
 
         public event EventHandler GameGameObjectAddGameObjectAssetVisible_EventHandler;
         public event EventHandler GameObjectAddAssetToCollection_EventHandler;
@@ -110,9 +135,15 @@ namespace Ursula.GameObjects.View
         {
         }
 
-        public GameObjectAddGameObjectAssetModel SetAddGameObjectAssetToCollection(string modelName, string gameObjectGroup, int gameObjectClass, string gameObjectSample)
+        public GameObjectAddGameObjectAssetModel SetCurrentAssetToCollection()
         {
-            SetModelName(modelName);
+            SetAddGameObjectAssetToCollection(ModelName, GameObjectGroup, GameObjectClass, GameObjectSample);
+            return this;
+        }
+
+        public GameObjectAddGameObjectAssetModel SetAddGameObjectAssetToCollection(string _modelName, string _gameObjectGroup, int _gameObjectClass, string _gameObjectSample)
+        {
+            SetModelName(_modelName);
 
             CopyGameObjectAsset();
 
@@ -129,9 +160,9 @@ namespace Ursula.GameObjects.View
             template = new GameObjectTemplate
                 (
                 directoryInfo.Name,
-                gameObjectGroup,
-                gameObjectClass,
-                gameObjectSample,
+                _gameObjectGroup,
+                _gameObjectClass,
+                _gameObjectSample,
                 sources,
                 Path.GetFileName(_gameObjectUserSourceData.GraphXmlPath),
                 Path.GetFileName(PreviewImageFilePathFrom)
@@ -211,26 +242,47 @@ namespace Ursula.GameObjects.View
             return this;
         }
 
-        public GameObjectAddGameObjectAssetModel SetEditAsset(GameObjectAssetInfo assetInfo)
+        public GameObjectAddGameObjectAssetModel SetGameObjectGroup(string value)
         {
-            this.assetInfo = assetInfo;
-            template = assetInfo.Template;
+            if (_gameObjectUserSourceData.GameObjectGroup == value) return this;
+            _gameObjectUserSourceData.SetGameObjectGroup(value);
+            return this;
+        }
 
-            SetModelName(assetInfo.Name);
-            SetModelPath(assetInfo.Template.Sources.Model3dFilePath);
-            SetGraphXmlPath(assetInfo.Template.GraphXmlPath);
-            SetPreviewImageFilePath(assetInfo.Template.PreviewImageFilePath);
-            SetDestPath(assetInfo.Template.Folder);
+        public GameObjectAddGameObjectAssetModel SetGameObjectClass(int value)
+        {
+            if (_gameObjectUserSourceData.GameObjectClass == value) return this;
+            _gameObjectUserSourceData.SetGameObjectClass(value);
+            return this;
+        }
 
-            SetGameObjectAddGameObjectAssetVisible(false, true);
+        public GameObjectAddGameObjectAssetModel SetGameObjectSample(string value)
+        {
+            if (_gameObjectUserSourceData.GameObjectSample == value) return this;
+            _gameObjectUserSourceData.SetGameObjectSample(value);
+            return this;
+        }
 
+        public GameObjectAddGameObjectAssetModel ClearAudios()
+        {
+            _gameObjectUserSourceData.AudiosFrom = new List<string>();
+            _gameObjectUserSourceData.AudiosTo = new List<string>();
+            return this;
+        }
+
+        public GameObjectAddGameObjectAssetModel ClearAnimations()
+        {
+            _gameObjectUserSourceData.AnimationsFrom = new List<string>();
+            _gameObjectUserSourceData.AnimationsTo = new List<string>();
             return this;
         }
 
         private void CopyGameObjectAsset()
         {
             MapManager.CreateDir(DestPath);
-            ModelLoader.CopyModel(ModelPath, DestPath);
+
+            if (Provider == GameObjectAssetsUserSource.LibId)
+                ModelLoader.CopyModel(ModelPath, DestPath);
 
             _gameObjectUserSourceData.AudiosTo = new List<string>();
             string pathAudio = DestPath + MapManager.PATHAUDIO;
@@ -239,11 +291,11 @@ namespace Ursula.GameObjects.View
             {
                 string pathFrom = _gameObjectUserSourceData.AudiosFrom[i];
                 string pathTo = $"{pathAudio}/{Path.GetFileName(_gameObjectUserSourceData.AudiosFrom[i])}";
-                if (File.Exists(pathFrom))
+                if (File.Exists(pathFrom) && pathFrom != pathTo)
                 {
-                    File.Copy(pathFrom, ProjectSettings.GlobalizePath(pathTo), true);
-                    _gameObjectUserSourceData.AudiosTo.Add($"{MapManager.PATHAUDIO}/{Path.GetFileName(pathTo)}");
+                    File.Copy(pathFrom, ProjectSettings.GlobalizePath(pathTo), true);               
                 }
+                _gameObjectUserSourceData.AudiosTo.Add($"{MapManager.PATHAUDIO}/{Path.GetFileName(pathTo)}");
             }
 
             _gameObjectUserSourceData.AnimationsTo = new List<string>();
@@ -254,11 +306,11 @@ namespace Ursula.GameObjects.View
                 string pathFrom = _gameObjectUserSourceData.AnimationsFrom[i];
                 string pathTo = $"{pathAnimations}/{Path.GetFileName(_gameObjectUserSourceData.AnimationsFrom[i])}";
 
-                if (File.Exists(pathFrom))
+                if (File.Exists(pathFrom) && pathFrom != pathTo)
                 {
                     File.Copy(pathFrom, ProjectSettings.GlobalizePath(pathTo), true);
-                    _gameObjectUserSourceData.AnimationsTo.Add($"{MapManager.PATHANIMATION}/{Path.GetFileName(pathTo)}");
                 }
+                _gameObjectUserSourceData.AnimationsTo.Add($"{MapManager.PATHANIMATION}/{Path.GetFileName(pathTo)}");
             }
 
             if (File.Exists(GraphXmlPathFrom)
